refactor: rearrange the files into layered architecture
